V1.0 for collections and threads
Note that for the AtomicInteger test (#3 under threads), I ended up changing the collection from ArrayList to CopyOnWriteArrayList.   The latter is thread safe and this resolved the intermittent results I was receiving.   Another approach would have been to do a lock on the entire Array List before adding items to it - while a valid approach, this would require some performance testing to see if one is better than the other.